+++ docs/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Tome to "Tools and Utilities" tab in creative mode (#42)
+
+### Fixed
+
+- Tome missing from creative mode and JEI search (#42)
+
 ## [1.10.1] - 2023-06-11
 
 ### Added